requirements.txt
fastapi
uvicorn
pytest
httpx
pydantic
email-validator
PyYAML